[refactor](mysql_table_writer)split into two parts of vectorized and row mode
